Make sidebar manage shortcuts
